Fix hardening when water flow into lava. Fix entity kill so that death animation have time to play before despawn. Pickup items even when in creative. Added mutton, etc. Blacklisted MiNET banned dude. Added vanilla teleport and enchant commands.

+++ src/MiNET/MiNET.BuilderBase/BuilderBaseItemFactory.cs
@@ -6,7 +6,7 @@ namespace MiNET.BuilderBase
 {
 	public class BuilderBaseItemFactory : ICustomItemFactory
 	{
-		public Item GetItem(short id, short metadata, byte count)
+		public Item GetItem(short id, short metadata, int count)
 		{
 			if (id == new BrushTool().Id)
 			{
